feedback improvements [7]
